+++ packages/components/src/ConfirmationModal/ConfirmationModal.tsx
@@ -5,6 +5,7 @@ import React, {
   useImperativeHandle,
   useReducer,
 } from "react";
+import { useOnKeyDown } from "@jobber/hooks";
 import { Modal } from "../Modal";
 import { Content } from "../Content";
 import { Markdown } from "../Markdown";
@@ -22,7 +23,6 @@ interface ConfirmationModalState {
 interface ConfirmOrCancelAction {
   type: "confirm" | "cancel";
 }
-
 interface DisplayAction {
   type: "display";
   title?: string;
@@ -108,6 +108,11 @@ interface BaseConfirmationModalProps {
   readonly cancelLabel?: string;
 
   /**
+   * Type (Work or destructive) for confirm button.
+   */
+  readonly variation?: "work" | "destructive";
+
+  /**
    * Callback for when the confirm button is pressed.
    */
   onConfirm?(): void;
@@ -148,6 +153,7 @@ export const ConfirmationModal = forwardRef(function ConfirmationModalInternal(
     onConfirm,
     onCancel,
     onRequestClose,
+    variation = "work",
   }: ConfirmationModalProps,
   ref: Ref<ConfirmationModalRef>,
 ) {
@@ -198,6 +204,8 @@ export const ConfirmationModal = forwardRef(function ConfirmationModalInternal(
     });
   }, [title, message, confirmLabel, cancelLabel, onConfirm, onCancel]);
 
+  useOnKeyDown(handleKeyboardShortcut, ["Escape", "Enter"]);
+
   return (
     <Modal
       title={state.title}
@@ -206,20 +214,45 @@ export const ConfirmationModal = forwardRef(function ConfirmationModalInternal(
       dismissible={false}
       primaryAction={{
         label: state.confirmLabel,
-        onClick: () => {
-          dispatch({ type: "confirm" });
-          onRequestClose && onRequestClose();
-        },
+        onClick: handleAction("confirm"),
+        variation: variation === "destructive" ? "destructive" : "work",
       }}
       secondaryAction={{
         label: state.cancelLabel,
-        onClick: () => {
-          dispatch({ type: "cancel" });
-          onRequestClose && onRequestClose();
-        },
+        onClick: handleAction("cancel"),
       }}
     >
       <Content>{state.message && <Markdown content={state.message} />}</Content>
     </Modal>
   );
+
+  function handleAction(type: "confirm" | "cancel") {
+    return () => {
+      dispatch({ type });
+      onRequestClose && onRequestClose();
+    };
+  }
+
+  function handleKeyboardShortcut(event: KeyboardEvent) {
+    const { metaKey, ctrlKey, key, target } = event;
+    if (!open) return;
+
+    const shouldTriggerShortcut =
+      target instanceof HTMLButtonElement ? metaKey || ctrlKey : true;
+
+    if (!shouldTriggerShortcut) return;
+
+    event.preventDefault();
+    event.stopPropagation();
+    switch (key) {
+      case "Enter": {
+        handleAction("confirm")();
+        break;
+      }
+      case "Escape": {
+        handleAction("cancel")();
+        break;
+      }
+    }
+  }
 });