+++ packages/hooks/CHANGELOG.md
@@ -8,6 +8,36 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.6.1...@jobber/hooks@1.7.0) (2021-09-23)
+
+
+### Features
+
+* **components:** Added keyboard shortcut to confirmation modal  ([#669](https://github.com/GetJobber/atlantis/issues/669)) ([f93456f](https://github.com/GetJobber/atlantis/commit/f93456fcf0bd08472e194282d059572ee1583fc3))
+
+
+
+
+
+## [1.6.1](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.6.0...@jobber/hooks@1.6.1) (2021-09-21)
+
+**Note:** Version bump only for package @jobber/hooks
+
+
+
+
+
+# [1.6.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.5.3...@jobber/hooks@1.6.0) (2021-09-20)
+
+
+### Features
+
+* **eslint:** Update prettier to latest version ([#667](https://github.com/GetJobber/atlantis/issues/667)) ([84e5895](https://github.com/GetJobber/atlantis/commit/84e589542c642b27acd2997af7200b86ba197b7e))
+
+
+
+
+
 ## [1.5.3](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.5.2...@jobber/hooks@1.5.3) (2021-09-10)
 
 **Note:** Version bump only for package @jobber/hooks